# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.FailureErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  types.FailureResponse:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  types.TaskCreateRequest:
    properties:
      assignee_id:
        type: string
      description:
        type: string
      status:
        description: todo, in_progress, done, cancelled
        type: string
      team_id:
        type: string
      title:
        type: string
    type: object
  types.TaskResponse:
    properties:
      assignee_id:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        type: string
      team_id:
        type: string
      title:
        type: string
    type: object
  types.TaskUpdateRequest:
    properties:
      assignee_id:
        type: string
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
info:
  contact: {}
  description: This is an API of auth-service
  title: Service task  Swagger
  version: "1.0"
paths:
...
  /tasks:
    get:
      description: ListTasks возвращает список задач с фильтрами и пагинацией.
      parameters:
      - description: team ID
        in: query
        name: team_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: assignee ID
        in: query
        name: assignee_id
        type: string
      - description: page
        in: query
        name: page
        type: string
      - description: page_size
        in: query
        name: page_size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: listTasks info
      tags:
      - task
    post:
      description: CreateTask создаёт задачу для команды (только член команды).
      parameters:
      - description: task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/types.TaskCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: createTask info
      tags:
      - task
  /tasks/{id}:
    put:
      description: UpdateTask обновляет задачу и пишет историю изменений.
      parameters:
      - description: task ID
        in: path
        name: id
        type: string
      - description: task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/types.TaskUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: updateTask info
      tags:
      - task
  /tasks/{id}/history:
    get:
      description: GetTaskHistory возвращает историю изменений задачи.
      parameters:
      - description: task ID
        in: path
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.TaskResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.FailureResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.FailureErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: getTaskHistory info
      tags:
      - task
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"